+++ plugins/cpu/z80-cpu/src/main/java/net/emustudio/plugins/cpu/zilogZ80/InitializerForZ80.java
@@ -42,7 +42,7 @@ public class InitializerForZ80 extends DefaultInitializer<EmulatorEngine> {
     }
 
     @Override
-    protected EmulatorEngine createEmulatorEngine(MemoryContext<Short> memory) {
+    protected EmulatorEngine createEmulatorEngine(MemoryContext<Byte> memory) {
         return new EmulatorEngine(memory, context);
     }
 
@@ -53,7 +53,7 @@ public class InitializerForZ80 extends DefaultInitializer<EmulatorEngine> {
     }
 
     @Override
-    protected Disassembler createDisassembler(MemoryContext<Short> memory) {
+    protected Disassembler createDisassembler(MemoryContext<Byte> memory) {
         return new DisassemblerImpl(memory, new DecoderImpl(memory));
     }
 }
